requirements.txt
gradio>=4.44.0,<6.0.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
rich>=13.0.0
scipy>=1.11.0
python-dotenv>=1.0.0
numpy>=1.24.0
aiohttp>=3.9.0
tabulate>=0.9.0
